+++ src/givr/cube.h
@@ -7,7 +7,7 @@ namespace givr {
     };
 
     std::vector<float> generate_vertices(cube const &s);
-    std::vector<std::uint16_t> generate_indices(cube const &s);
+    std::vector<std::uint32_t> generate_indices(cube const &s);
 
     // TODO: implement the uv and normal functions.
 }; // end namespace givr
